move logger abstraction to a separate file
