+++ v2/logging.go
@@ -1,0 +1,36 @@
+package shuttle
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"time"
+)
+
+type Logger interface {
+	Info(s string)
+	Warn(s string)
+	Error(s string)
+}
+
+var getLogger = func(_ context.Context) Logger { return &printLogger{} }
+
+type printLogger struct{}
+
+func (l *printLogger) Info(s string) {
+	fmt.Println(append(append([]any{}, "INFO - ", time.Now().UTC(), " - "), s)...)
+}
+
+func (l *printLogger) Warn(s string) {
+	fmt.Println(append(append([]any{}, "WARN - ", time.Now().UTC(), " - "), s)...)
+}
+
+func (l *printLogger) Error(s string) {
+	fmt.Println(append(append([]any{}, "ERROR - ", time.Now().UTC(), " - "), s)...)
+}
+
+func log(ctx context.Context, a ...any) {
+	if os.Getenv("GOSHUTTLE_LOG") == "ALL" {
+		getLogger(ctx).Info(fmt.Sprint(append(append([]any{}, time.Now().UTC(), " - "), a...)...))
+	}
+}
