+++ v2/processor.go
@@ -4,14 +4,14 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"os"
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus"
-	"github.com/Azure/go-shuttle/v2/metrics"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/trace"
+
+	"github.com/Azure/go-shuttle/v2/metrics"
 )
 
 type Receiver interface {
@@ -217,33 +217,5 @@ func (plr *peekLockRenewer) startPeriodicRenewal(ctx context.Context, message *a
 			}
 			alive = false
 		}
-	}
-}
-
-type Logger interface {
-	Info(s string)
-	Warn(s string)
-	Error(s string)
-}
-
-var getLogger = func(_ context.Context) Logger { return &printLogger{} }
-
-type printLogger struct{}
-
-func (l *printLogger) Info(s string) {
-	fmt.Println(append(append([]any{}, "INFO - ", time.Now().UTC(), " - "), s)...)
-}
-
-func (l *printLogger) Warn(s string) {
-	fmt.Println(append(append([]any{}, "WARN - ", time.Now().UTC(), " - "), s)...)
-}
-
-func (l *printLogger) Error(s string) {
-	fmt.Println(append(append([]any{}, "ERROR - ", time.Now().UTC(), " - "), s)...)
-}
-
-func log(ctx context.Context, a ...any) {
-	if os.Getenv("GOSHUTTLE_LOG") == "ALL" {
-		getLogger(ctx).Info(fmt.Sprint(append(append([]any{}, time.Now().UTC(), " - "), a...)...))
 	}
 }
